increase mode-line contrast for multiple window
I prefer the look of the less contrasty mode-line, so use that when
there is only one window. If there are multiple windows, we need
modelines that

  1. are distinct enough to see window structure at a glance
  2. differentiate the active mode-line enough

That means we need two obviously different colors that still look
good. Before this commit, the inactive style line was too light so it
was hard to see all the windows. This commit makes the inactive modeline
the same color as the original modeline. The active modeline is
darker. But the darker line looks bad with just one buffer, so we
preserve the original look when there is just one buffer.