add mode-specific company (autocompletion) handling
company mode autocompletion does not configure itself! backends for
each mode need to be manually set. the recommended approach is to add
mode-specific backends to the front of the default backends list, as
in:

    (setq (make-local-variable 'company-backends)
      (append 'company-anaconda company-backends))

This buffer-local variable is then set in the mode hook, so it applies
to all buffers using the mode:

    (add-hook 'python-mode-hook
      (lambda ()
        (setq (make-local-variable 'company-backends ...))))

which can get a little verbose, so I defined a macro to make life
easier:

   (jyh-company-for-mode 'python-mode-hook company-anaconda)