use environment/path set automatically by pyenv/nodenv
from within emacs, external commands are found via two mechanisms:
process-environment and exec-path:

1) `process-environment` is a set of environment variables like PATH
that would exist in a normal shell. Commands called
via (shell-command) operate within the `process-environment`.

2) `exec-path` is a list of directories to search for executables
when (create-process) and its friends are called.

In both cases, given that pyenv/rbenv/nodenv's shims work operate
based on the working directory, all the work in emacs is just to make
sure that 1) the path in both process-environment and exec-path are
set to find the right shim, and 2) that the command is run from the
file's containing directory (which happens by default in emacs
already).